Move stage instances related modules to channels module
