+++ src/helpers/channels/get_stage_instance.ts
@@ -1,5 +1,5 @@
 import { rest } from "../../rest/rest.ts";
-import { StageInstance } from "../../types/misc/stage_instance.ts";
+import { StageInstance } from "../../types/channels/stage_instance.ts";
 import { endpoints } from "../../util/constants.ts";
 
 /** Gets the stage instance associated with the Stage channel, if it exists. */
