+++ src/types/mod.ts
@@ -231,4 +231,4 @@ export * from "./webhooks/execute_webhook.ts";
 export * from "./webhooks/modify_webhook.ts";
 export * from "./webhooks/webhook.ts";
 export * from "./webhooks/webhooks_update.ts";
-export * from "./misc/stage_instance.ts";
+export * from "./channels/stage_instance.ts";
